[ROM] Add collection specs
